Move askpass code to it's own crate
