+++ crates/askpass/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "askpass"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/askpass.rs"
+
+[dependencies]
+anyhow.workspace = true
+futures.workspace = true
+gpui.workspace = true
+smol.workspace = true
+tempfile.workspace = true
+util.workspace = true
+which.workspace = true
