+++ crates/remote/Cargo.toml
@@ -19,6 +19,7 @@ test-support = ["fs/test-support"]
 
 [dependencies]
 anyhow.workspace = true
+askpass.workspace = true
 async-trait.workspace = true
 collections.workspace = true
 fs.workspace = true
@@ -26,9 +27,10 @@ futures.workspace = true
 gpui.workspace = true
 itertools.workspace = true
 log.workspace = true
-paths.workspace = true
 parking_lot.workspace = true
+paths.workspace = true
 prost.workspace = true
+release_channel.workspace = true
 rpc = { workspace = true, features = ["gpui"] }
 schemars.workspace =  true
 serde.workspace = true
@@ -38,7 +40,6 @@ smol.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
 util.workspace = true
-release_channel.workspace = true
 which.workspace = true
 
 [dev-dependencies]
