+++ crates/remote/src/ssh_session.rs
@@ -1454,6 +1454,8 @@ impl SshRemoteConnection {
         delegate: Arc<dyn SshClientDelegate>,
         cx: &mut AsyncApp,
     ) -> Result<Self> {
+        use askpass::AskPassResult;
+
         delegate.set_status(Some("Connecting"), cx);
 
         let url = connection_options.ssh_url();
@@ -1462,7 +1464,7 @@ impl SshRemoteConnection {
             .prefix("zed-ssh-session")
             .tempdir()?;
 
-        let mut askpass = AskPassSession::new(&temp_dir, cx, {
+        let mut askpass = askpass::AskPassSession::new(&temp_dir, cx, {
             let delegate = delegate.clone();
             move |prompt, cx| delegate.ask_password(prompt, cx)
         })
@@ -2509,137 +2511,5 @@ mod fake {
         }
 
         fn set_status(&self, _: Option<&str>, _: &mut AsyncApp) {}
-    }
-}
-
-struct AskPassSession {
-    script_path: PathBuf,
-    _askpass_task: Task<()>,
-    askpass_opened_rx: Option<oneshot::Receiver<()>>,
-    askpass_kill_master_rx: Option<oneshot::Receiver<()>>,
-}
-
-#[derive(PartialEq, Eq)]
-pub enum AskPassResult {
-    CancelledByUser,
-    Timedout,
-}
-
-impl AskPassSession {
-    /// This will create a new AskPassSession.
-    /// You must retain this session until the master process exits.
-    #[must_use]
-    pub async fn new(
-        temp_dir: &TempDir,
-        cx: &mut AsyncApp,
-        password_prompt: impl Fn(String, &mut AsyncApp) -> oneshot::Receiver<Result<String>>
-            + Send
-            + Sync
-            + 'static,
-    ) -> Result<Self> {
-        use futures::AsyncWriteExt as _;
-        use futures::{io::BufReader, AsyncBufReadExt as _};
-        use smol::{fs::unix::PermissionsExt as _, net::unix::UnixListener};
-        use util::ResultExt as _;
-
-        let askpass_socket = temp_dir.path().join("askpass.sock");
-        let (askpass_opened_tx, askpass_opened_rx) = oneshot::channel::<()>();
-        let listener =
-            UnixListener::bind(&askpass_socket).context("failed to create askpass socket")?;
-
-        let (askpass_kill_master_tx, askpass_kill_master_rx) = oneshot::channel::<()>();
-        let mut kill_tx = Some(askpass_kill_master_tx);
-
-        let askpass_task = cx.spawn(|mut cx| async move {
-            let mut askpass_opened_tx = Some(askpass_opened_tx);
-
-            while let Ok((mut stream, _)) = listener.accept().await {
-                if let Some(askpass_opened_tx) = askpass_opened_tx.take() {
-                    askpass_opened_tx.send(()).ok();
-                }
-                let mut buffer = Vec::new();
-                let mut reader = BufReader::new(&mut stream);
-                if reader.read_until(b'\0', &mut buffer).await.is_err() {
-                    buffer.clear();
-                }
-                let prompt = String::from_utf8_lossy(&buffer);
-                if let Some(password) = password_prompt(prompt.to_string(), &mut cx)
-                    .await
-                    .context("failed to get askpass password")
-                    .and_then(|p| p)
-                    .log_err()
-                {
-                    stream.write_all(password.as_bytes()).await.log_err();
-                } else {
-                    if let Some(kill_tx) = kill_tx.take() {
-                        kill_tx.send(()).log_err();
-                    }
-                    // note: we expect the caller to drop this task when it's done.
-                    // We need to keep the stream open until the caller is done to avoid
-                    // spurious errors from ssh.
-                    std::future::pending::<()>().await;
-                    drop(stream);
-                }
-            }
-        });
-
-        anyhow::ensure!(
-            which::which("nc").is_ok(),
-            "Cannot find `nc` command (netcat), which is required to connect over SSH."
-        );
-
-        // Create an askpass script that communicates back to this process.
-        let askpass_script = format!(
-            "{shebang}\n{print_args} | {nc} -U {askpass_socket} 2> /dev/null \n",
-            // on macOS `brew install netcat` provides the GNU netcat implementation
-            // which does not support -U.
-            nc = if cfg!(target_os = "macos") {
-                "/usr/bin/nc"
-            } else {
-                "nc"
-            },
-            askpass_socket = askpass_socket.display(),
-            print_args = "printf '%s\\0' \"$@\"",
-            shebang = "#!/bin/sh",
-        );
-        let askpass_script_path = temp_dir.path().join("askpass.sh");
-        fs::write(&askpass_script_path, askpass_script).await?;
-        fs::set_permissions(&askpass_script_path, std::fs::Permissions::from_mode(0o755)).await?;
-
-        Ok(Self {
-            script_path: askpass_script_path,
-            _askpass_task: askpass_task,
-            askpass_kill_master_rx: Some(askpass_kill_master_rx),
-            askpass_opened_rx: Some(askpass_opened_rx),
-        })
-    }
-
-    fn script_path(&self) -> &Path {
-        &self.script_path
-    }
-
-    // This will run the askpass task forever, resolving as many authentication requests as needed.
-    // The caller is responsible for examining the result of their own commands and cancelling this
-    // future when this is no longer needed. Note that this can only be called once, but due to the
-    // drop order this takes an &mut, so you can `drop()` it after you're done with the master process.
-    async fn run(&mut self) -> AskPassResult {
-        let connection_timeout = Duration::from_secs(10);
-        let askpass_opened_rx = self.askpass_opened_rx.take().expect("Only call run once");
-        let askpass_kill_master_rx = self
-            .askpass_kill_master_rx
-            .take()
-            .expect("Only call run once");
-
-        select_biased! {
-            _ = askpass_opened_rx.fuse() => {
-                // Note: this await can only resolve after we are dropped.
-                askpass_kill_master_rx.await.ok();
-                return AskPassResult::CancelledByUser
-            }
-
-            _ = futures::FutureExt::fuse(smol::Timer::after(connection_timeout)) => {
-                return AskPassResult::Timedout
-            }
-        }
     }
 }
